requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.1.31
charset-normalizer==3.4.1
datasets==3.5.0
dill==0.3.8
filelock==3.18.0
frozenlist==1.6.0
fsspec==2024.12.0
gdown==5.2.0
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
networkx==3.4.2
numpy==2.2.5
packaging==25.0
pandas==2.2.3
propcache==0.3.1
psutil==7.0.0
pyarrow==19.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
salad-cloud-sdk==0.9.0a12
scikit-learn==1.6.1
scipy==1.15.2
seqeval==1.2.2
setuptools==79.0.1
six==1.17.0
soupsieve==2.7
sympy==1.13.3
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wheel==0.45.1
xxhash==3.5.0
yarl==1.20.0
